# docs/swagger.yaml
definitions:
...
  controller.LoginDto:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controller.UserDto:
    properties:
      avatar:
        type: string
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - avatar
    - email
    - password
    - phone
    - username
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      success:
        type: boolean
      timestamp:
        type: integer
    type: object
info:
  contact: {}
  description: 博客主要包含用户、文章和分类
  license:
    name: Apache 2.0
  title: Gin博客项目接口文档
  version: "1.0"
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录，返回token
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 登录参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/controller.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户登录
      tags:
      - 用户接口
...
  /api/user/create:
    post:
      consumes:
      - application/json
      description: 添加用户，参数为username和password
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/controller.UserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 添加用户
      tags:
      - 用户接口
  /api/user/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 通过用户id删除用户
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除用户
      tags:
      - 用户接口
  /api/user/detail/{id}:
    get:
      consumes:
      - application/json
      description: 通过用户id获取用户详情
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取用户详情页
      tags:
      - 用户接口
  /api/user/enable/{id}:
    get:
      consumes:
      - application/json
      description: 通过用户id禁用用户
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 激活用户
      tags:
      - 用户接口
  /api/user/list:
    get:
      consumes:
      - application/json
      description: 通过用户username获取用户列表数据
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户username
        in: query
        name: username
        type: string
      - description: 当前第X页
        in: query
        name: page_num
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 搜索用户列表
      tags:
      - 用户接口
  /api/user/reset/{id}:
    get:
      consumes:
      - application/json
      description: 通过用户id重置登录密码为123456
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 重置密码
      tags:
      - 用户接口
swagger: "2.0"